hook up deserializer up to some test code, dont try to insert anything yet, just deserialize
