+++ relational_db_builder/src/main.rs
@@ -1,10 +1,13 @@
 mod packument;
 
+use serde_json::{Map, Value};
 use postgres_db::DbConnection;
 use postgres_db::internal_state;
 use postgres_db::change_log;
 use postgres_db::change_log::Change;
 use utils::check_no_concurrent_processes;
+
+use utils::RemoveInto;
 
 const PAGE_SIZE: i64 = 1024;
 
@@ -44,5 +47,20 @@ fn main() {
 }
 
 fn process_change(conn: &DbConnection, c: Change) {
+    let mut change_json = serde_json::from_value::<Map<String, Value>>(c.raw_json).unwrap();
+    let del = change_json.remove_key_unwrap_type::<bool>("deleted").unwrap();
+
+    if del {
+        panic!("Don't yet know how to handle deleted things");
+    }
+
+    let package_name = change_json.remove_key_unwrap_type::<String>("id").unwrap();
+
+    if package_name == "_design/app" || package_name == "_design/scratch" {
+        return
+    }
+
+    let doc = change_json.remove_key_unwrap_type::<Map<String, Value>>("doc").unwrap();
+    let packument = packument::deserialize::deserialize_packument_blob(doc).unwrap();
 
 }
